pyproject.toml
[project]
name = "pyxeled"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2"
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]

[build-system]
requires = ["maturin>=1,<2"]
build-backend = "maturin"
